requirements.txt
streamlit==1.57.0
tensorflow==2.21.0
numpy==2.4.4
pillow==12.2.0
